\name{RcppCImg-package}
\alias{RcppCImg-package}
\alias{RcppCImg}
\docType{package}
\title{
Bindings for the CImg image manipulation library
}
\description{
The package exposes classes from the CImg C++ library 
to the R level via Rcpp modules
}
\details{
\tabular{ll}{
Package: \tab RcppCImg\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2010-11-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois <romain@r-enthusiasts.com>

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\references{
    CImg: \url{http://cimg.sourceforge.net/}
}
\keyword{ package }

